Import of RURBAN/B-C-1.45_06 from CPAN.
gitpan-cpan-distribution: B-C
gitpan-cpan-version:      1.45_06
gitpan-cpan-path:         RURBAN/B-C-1.45_06.tar.gz
gitpan-cpan-author:       RURBAN
gitpan-cpan-maturity:     developer

+++ t/C-COMPILED/CORE--op--readline.t
@@ -47,7 +47,7 @@ pass( $taint ? "Taint mode!" : "Not in taint mode" );
 ( my $bin_file = $file_to_test ) =~ s/\.t$/.bin/;
 unlink $bin_file, $c_file;
 
-my $PERL = $^X;
+my $PERL = $^X =~ m/\s/ ? qq{"$^X"} : $^X;
 
 my $check = `$PERL -c $taint '$file_to_test' 2>&1`;
 like( $check, qr/syntax OK/, "$PERL -c $taint $file_to_test" );
@@ -58,9 +58,22 @@ my %SIGNALS = qw( 11 SEGV 6 SIGABRT 1 SIGHUP 13 SIGPIPE);
 $SIGNALS{0} = '';
 
 foreach my $optimization (@optimizations) {
-  TODO: SKIP: {
+TODO: {
+  SKIP: {
         local $TODO = $todo if ( $todo =~ /B::C Fails to generate c code/ );
         local $ENV{BC_OPT} = $optimization;
+
+        my $b = $optimization; # protect against parallel test name clashes
+        #$b =~ s/-(D.*|f.*|v),//g;
+        #$b =~ s/-/_/g;
+        #$b =~ s/[, ]//g;
+        #$b =~ s/_O0$//;
+        #$b = lc($b);
+        $b = ''; # need to check $0 diagnostics
+        ( $c_file   = $file_to_test ) =~ s/\.t$/$b.c/;
+        $b = '.bin'; # need to check $0 diagnostics
+        ( $bin_file = $file_to_test ) =~ s/\.t$/$b/;
+        unlink $bin_file, $c_file;
 
         # Generate the C code at $optimization level
         my $cmd = "$PERL $taint -Iblib/arch -Iblib/lib -MO=-qq,C,$optimization,-o$c_file $file_to_test 2>&1";
@@ -87,7 +100,7 @@ foreach my $optimization (@optimizations) {
         ok( -x $bin_file, "$bin_file is compiled and ready to run." );
 
         if ( !-x $bin_file ) {
-            unlink $c_file, $bin_file;
+            unlink $c_file, $bin_file unless $ENV{BC_DEVELOPING};
             skip( "Can't test further due to failure to create a binary file.", 8 );
         }
 
@@ -137,8 +150,8 @@ foreach my $optimization (@optimizations) {
         ok( !scalar @{ $parser->{failed} }, "Test results:" );
         print "    $_\n" foreach ( split( "\n", $out ) );
 
-        if (!ok( !scalar @{ $parser->{failed} }, "No test failures" )) {
-          note( "Failed tests: " . join( ", ", @{ $parser->{failed} } ) );
+        if (!ok( !scalar @{ $parser->{failed} }, "No test failures $optimization" )) {
+          note( "Failed $optimization tests: " . join( ", ", @{ $parser->{failed} } ) );
           $ENV{BC_DEVELOPING} = 1; # keep temp files
         }
 
@@ -151,12 +164,13 @@ foreach my $optimization (@optimizations) {
           $ENV{BC_DEVELOPING} = 1; # keep temp files
         }
 
-        local $TODO = "tests unexpectedly passing";
-        if (!ok( !scalar @{ $parser->{todo_passed} }, "No TODO tests passed" )) {
+        local $TODO = "tests unexpectedly passing" if scalar @{ $parser->{todo_passed} };
+        if (!ok( !scalar @{ $parser->{todo_passed} }, "No TODO tests passed $optimization" )) {
           note( "TODO Passed: " . join( ", ", @{ $parser->{todo_passed} } ) );
           $ENV{BC_DEVELOPING} = 1; # keep temp files
         }
         $TODO = '';
     }
+  }
+  unlink $bin_file, $c_file unless $ENV{BC_DEVELOPING};
 }
-unlink $bin_file, $c_file unless $ENV{BC_DEVELOPING};
